Edit Admin

+++ ToplearnBlogProject.Application/Services/Repositories/AdminRepository.cs
@@ -33,34 +33,68 @@ namespace ToplearnBlogProject.Application.Services.Repositories
         }
         public async Task<List<Admin>> GetAll()
         {
-            //return await Task.FromResult(_dbContext.Admins.ToList());
-            return _dbContext.Admins.Include(p => p.Role).ToList();
+            var data = _dbContext.Admins.Include(p => p.Role)
+                .Select(p => new Admin
+                {
+                    Role = p.Role,
+                    BlogCount = p.BlogCount,
+                    Id = p.Id,
+                    Name = p.Name,
+                    LastName = p.LastName,
+                    Email = p.Email,
+                    IsActive = p.IsActive,
+                    Phone = p.Phone,
+                    CreatedAt = p.CreatedAt,
+                    RoleId = p.RoleId,
+                    Avatar = p.Avatar != null ? string.Concat(nameof(Admin).ToLower(), "/", p.Avatar) : null
+                }).ToList();
+
+            return data;
+            //return _dbContext.Admins.Include(p => p.Role).ToList();
         }
 
-        public async Task<bool> Remove(int id)
+        public async Task<string> Remove(int id)
         {
-            Admin admin = new Admin
+            var admin = _dbContext.Admins.FirstOrDefault(p => p.Id == id);
+            if (admin != null)
             {
-                Id = id
-            };
-            _dbContext.Admins.Remove(admin);
+                var avatar = admin.Avatar;
+                _dbContext.Admins.Remove(admin);
 
-            try
-            {
-                await _dbContext.SaveChangesAsync();
-                return true;
+                try
+                {
+                    await _dbContext.SaveChangesAsync();
+                    return avatar;
+                }
+                catch (Exception e)
+                {
+                    throw new Exception(e.Message);
+                }
             }
-            catch (Exception e)
-            {
-                throw new Exception(e.Message);
-                throw;
-            }
+            throw new Exception("Admin Not Found");
         }
         public async Task<Admin> FindById(int id)
         {
             //return await Task.FromResult(_dbContext.Admins.FirstOrDefault(p => p.Id == id));
 
-            var result = _dbContext.Admins.Where(p => p.Id == id).Include(p => p.Role).FirstOrDefault();
+            var result = _dbContext.Admins
+                .Where(p => p.Id == id)
+                .Include(p => p.Role)
+                .Select(p => new Admin
+                {
+                    Role = p.Role,
+                    BlogCount = p.BlogCount,
+                    Id = p.Id,
+                    Name = p.Name,
+                    LastName = p.LastName,
+                    Email = p.Email,
+                    IsActive = p.IsActive,
+                    Phone = p.Phone,
+                    CreatedAt = p.CreatedAt,
+                    RoleId = p.RoleId,
+                    Avatar = p.Avatar != null ? string.Concat(nameof(Admin).ToLower(), "/", p.Avatar) : null
+                })
+                .FirstOrDefault();
 
             if (result != null)
             {
@@ -70,6 +104,7 @@ namespace ToplearnBlogProject.Application.Services.Repositories
         }
         public async Task<bool> Update(Admin data)
         {
+            //var admin = await _dbContext.Admins.FirstOrDefaultAsync(p => p.Id == data.Id);
             _dbContext.Admins.Update(data);
 
             try
@@ -89,7 +124,7 @@ namespace ToplearnBlogProject.Application.Services.Repositories
         Task<int> Create(Admin model);
         Task<Admin> FindById(int id);
         Task<List<Admin>> GetAll();
-        Task<bool> Remove(int id);
+        Task<string> Remove(int id);
         Task<bool> Update(Admin data);
     }
 }
